brief: modified exportDialog
